+++ index.js
@@ -1,14 +1,4 @@
 const fp = require("fastify-plugin");
-
-const FINGERPRINT_HEADERS = [
-  "connection",
-  "host",
-  "user-agent",
-  "accept",
-  "accept-encoding",
-  "accept-language",
-];
-
 const defaultHashFn = require("./src/hash");
 const filterHeaders = require("./src/filter-headers");
 
@@ -17,13 +7,18 @@ module.exports = fp(
     fastify,
     {
       acceptHeaders = undefined,
+      extendHeaders = undefined,
       requestKey = "fingerprint",
       hashFn = defaultHashFn,
     } = {},
     next
   ) => {
     const preValidation = async (request) => {
-      const values = filterHeaders(request.headers, acceptHeaders);
+      const values = filterHeaders(
+        request.headers,
+        acceptHeaders,
+        extendHeaders
+      );
       request[requestKey] = hashFn(values);
     };