+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### v1.1.0
 
-- [ ] Extend the default headers with a custom list of additional headers
+- [x] Extend the default headers with a custom list of additional headers
 - [ ] Extend the fingerprint using a list of cookies names to be included in the hash
 - [ ] Set a number of randomic cookie with randomic values, then use only a few of those to decorate the fingerprint